% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remoteSpeciesDataFunctions.R
\name{avisAllSpecies}
\alias{avisAllSpecies}
\title{avisAllSpecies}
\usage{
avisAllSpecies()
}
\value{
returns a vector
}
\description{
Returns a vector with the ids of the species in Proyecto AVIS
}
\note{
This functions does not allow arguments
}
\examples{
\dontrun{
avisAllSpecies()
}
}

