% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setupFunctions.R
\name{avisSetup}
\alias{avisSetup}
\title{avisSetup}
\usage{
avisSetup (...)
}
\arguments{
\item{...}{Package settings parameters. Available params: verbose = TRUE/FALSE}
}
\description{
Sets up settings that apply to the behabiour of the package
Allow users to turn off the information messages of the functions.
}
\examples{
\dontrun{
avisSetup(verbose=FALSE)
}
}

