% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remoteSpeciesDataFunctions.R
\name{avisSpeciesId}
\alias{avisSpeciesId}
\title{avisSpeciesId}
\usage{
avisSpeciesId(nameraw)
}
\arguments{
\item{nameraw}{scientific name of the species (e.g. "Pica pica")}
}
\value{
an integer
}
\description{
Returns the id of the selected species
}
\examples{
\dontrun{
avisSpeciesId("Pica pica")
}
}

