% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canarias}
\alias{canarias}
\title{A physical map of the Canary Islands}
\format{tif image}
\source{
http://www.openstreetmap.org/
}
\description{
A tif image downloaded from http://www.openstreetmap.org/ using the R- library OpenStreetMap
}
\keyword{datasets}

