% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ravis_shape_spain}
\alias{ravis_shape_spain}
\title{A Spanish administrative map}
\format{shapefile}
\source{
http://www.diva-gis.org/
}
\description{
A shapefile downloaded from  http://www.diva-gis.org/
}
\keyword{datasets}

