% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ravisUTMLatLong}
\alias{ravisUTMLatLong}
\title{UTM-Latlong}
\format{matrix}
\description{
Geographic coordinates (lat-long) of the centroids of the Spanish UTM squares
}
\keyword{datasets}

