\name{avisSetup}
\alias{avisSetup}
\title{avisSetup}
\usage{
avisSetup (...)
}
\arguments{
  \item{verbose}{TRUE / FALSE}
}
\description{
Sets up settings that apply to the behabiour of the package
Allow users to turn off the information messages of the
functions.
}
\examples{
\dontrun{
avisSetup(verbose=FALSE)
}
}

