\name{avisSpeciesId}
\alias{avisSpeciesId}
\title{avisSpeciesId}
\usage{
avisSpeciesId(nameraw)
}
\arguments{
  \item{namerow}{scientific name of the species (e.g. "Pica
  pica")}
}
\value{
an integer
}
\description{
Returns the id of the selected species
}
\examples{
\dontrun{
avisSpeciesId("Pica pica")
}
}

